requirements.txt
torch==2.1.2
Pillow==10.1.0
pillow-heif==0.14.0
salesforce-lavis==1.0.2
similarities==1.1.6
tqdm==4.66.1
jupyter
